requirements.txt
streamlit==1.32.0
anthropic==0.7.4
python-dotenv==1.0.1
pypdf2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
httpx>=0.24.1
